Reorganizando projeto de com base na arquitetura hexagonal #12
